Fix signal reconciliation ignoring content changes for flat signals
Flat corpus signal tables (no BLOB) lacked a data_hash column, so
reconcile_corpus_signals fell back to existence-only checking — signals
were never updated once created, even when their content changed. This
caused deploy-ready signals to retain stale 400+ char deploy paths
after the elide_featuring fix, preventing deployment of tracks with
many featuring artists.

Changes:
- Add data_hash column to all 13 flat corpus signal tables
- Add content_hash_value() to SignalContentHash trait for consistent
  hash computation between insert() and reconciliation
- Make TypedSignalWrite::content_hash() delegate to concrete type
  without enum discriminant (was causing hash mismatches)
- Fix blob signal insert() to use content_hash_value() instead of
  compute_blob_hash() (FNV-1a vs SipHash mismatch caused blob signals
  to always rewrite — correct but wasteful)
- Remove existence-only fallback from reconcile_corpus_signals

Signal tables are Computed type, so the schema reconciler will
DROP+CREATE them on first startup, triggering full recomputation.

+++ src/meta/computations/analysis/tags.rs
@@ -634,7 +634,7 @@ pub fn execute_detect_compound_tags_for_inode(
         path: corpus_path,
         compounds,
     });
-    let new_hash = signal.content_hash() as i64;
+    let new_hash = signal.content_hash();
     let existing_hash = read_only_db.corpus_signal_inode_hash::<CompoundTagSignal>(inode);
     if existing_hash != Some(new_hash) {
         sender.write_typed_signal(signal, witness);
